minor tweaks 23.0

+++ app/models/artist_notification.rb
@@ -15,8 +15,8 @@ class ArtistNotification < ActiveRecord::Base
   validates :notice_type, presence: true
 
   def self.notification_cleanup
-    ArtistNotification.where(read: true).where("updated_at < ?", 3.days.ago).find_each do |notification|
-      notification.destroy
+    ArtistNotification.where(read: true).find_each do |notification|
+      notification.destroy unless notification.updated_at < 3.days.ago
     end
   end
